Rename `merlin.graph` to `merlin.dag`
