+++ merlin/dag/ops/identity.py
@@ -15,31 +15,16 @@
 #
 
 from merlin.core.dispatch import DataFrameType
-from merlin.graph.base_operator import BaseOperator
-from merlin.graph.selector import ColumnSelector
-from merlin.schema import Schema
+from merlin.dag.base_operator import BaseOperator
+from merlin.dag.selector import ColumnSelector
 
 
-class SubsetColumns(BaseOperator):
-    def __init__(self, label=None):
-        self._label = label or self.__class__.__name__
-        super().__init__()
-
-    def compute_input_schema(
-        self,
-        root_schema: Schema,
-        parents_schema: Schema,
-        deps_schema: Schema,
-        selector: ColumnSelector,
-    ) -> Schema:
-        return parents_schema - deps_schema
-
+class Identity(BaseOperator):
     def transform(self, col_selector: ColumnSelector, df: DataFrameType) -> DataFrameType:
         """Simply returns the selected output columns from the input dataframe
 
-        The main functionality of this operator has to do with computing the schemas
-        for `-` nodes in the Workflow graph, so very little has to happen in the
-        `transform` method.
+        The main functionality of this operator has to do with building the
+        Workflow graph, so very little has to happen in the `transform` method.
 
         Parameters
         -----------
@@ -54,7 +39,3 @@ class SubsetColumns(BaseOperator):
             Returns a transformed dataframe for this operator
         """
         return super()._get_columns(df, col_selector)
-
-    @property
-    def label(self) -> str:
-        return self._label
