+++ merlin/dag/ops/selection.py
@@ -18,8 +18,8 @@ from __future__ import annotations
 import logging
 
 from merlin.core.dispatch import DataFrameType
-from merlin.graph.base_operator import BaseOperator
-from merlin.graph.selector import ColumnSelector
+from merlin.dag.base_operator import BaseOperator
+from merlin.dag.selector import ColumnSelector
 from merlin.schema import Schema
 
 LOG = logging.getLogger("SelectionOp")
